pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "Cython>=3.0", "numpy>=1.22"]
build-backend = "setuptools.build_meta"

[project]
name = "shinier"
version = "0.1.0"
description = "Shinier: Scientific Histogram Intensity Normalization and Image Equalization in R/G/B."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Nicolas Dupuis-Roy", email = "nicolas.dupuis.roy@umontreal.ca" }
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.22,<2.1",
    "Pillow>=9.0.0",
    "tqdm>=4.60",
    "pydantic>=2.12",
]
[project.optional-dependencies]
dev = [
    "matplotlib>=3.8.0",
    "pytest>=8.3.0",
    "setuptools>=44.1.0",
    "scikit-image>=0.24.0",
    "colour-science >= 0.4.0",
    "sqlite >= 3.50.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["shinier*"]

[[tool.setuptools.ext-modules]]
name = "shinier._cconvolve"
sources = ["src/shinier/_cconvolve.pyx"]
language = "c++"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests/unit_tests",
    "tests/validation_tests"
]
markers = [
    "unit_tests: default functional",
    "validation_tests: full combinatorial sweep (slow)"
]